Added randomized testing to the cdl unit test class.
I avoided creating a generalized testing framework (as was indicated on the
readme). To be honest it won't be practical to essentially recreate unit tests.

I'll go the realistic route and keep at it as long as I can with the hacky
sort-of unit tests that I have now. As soon as things get too unweildy then I'll
want to switch over to gtest myself but for now that's avoidable and I'd rather
code up more intersting things.