Bugfix with another ABA problem.
First, there was one hidden issue with string parsing for vector equality
in testing. That was taken care of.

There's another issue which results in thread starvation because the queue gets
sliced in half by an ABA issue that popped up on the tail_ side of the shared_queue.

GCC was being finnicky and this turned out to be a known bug, so I had to add
a macro-based test.

Switching to shared_ptrs everywhere also involved changing the queue destructor,
since the default recursive mechanism is insufficient for large lists
(causes stack overflow, or segfaults).