Fleshed out hazard_ptr a bit more.
Still does nothing, but I added some internal hazard_record class and it's
clear what the control flow looks like. The main part of the code that actually
does stuff has yet to come.

With the docs ready, I have updated the README. This included setting a "reach"
goal to a lower priority.