Added multithreaded testing for hazard pointers.
Unfortunately, the lack of support in my version of gcc (4.8), which is pretty
modern, for thread_local destructors implies that there is currently a memory
leak whenever one uses hazard pointers.

The tests all pass currently but they fail a valgrind check. I'll submit
a patch which fixes the issue for noncompliant compilers.