Fixed memory leak for hazard pointers.
Turns out gcc really was supporting thread-local storage all along. The issue
was that I was blindly following the algorithm that was oh-so-elegantly described
in the hazard pointer article I was using.

Turns out that the article left a blind eye to the following issue:

The writer thread marks something for deletion, which is being used by a reader.
Then the writer thread exits, before it can delete its scheduled deltion. This is
an issue because then the writer thread exits with its "remnanats" of the retired
list still not taken care of.

There was also a bug (that had not surfaced yet) that involved using the rlist more
than once. This was fixed.

With the global collector, there are no memory leaks.