Fixed laggy writers issue.
There was another conceptual problem with the articles: the number of hazard
pointers in use never went down. This mean that writers would never actually
do a deletion scan because they would always wait until there are 5/4 as many
deactivated hazard records in their thread-local reserves. However, this can't
possible work because the writer can only have as many local retired HPs as there
ever were!

I think what was actually meant was the number of <active> hazard pointers.

To represent this number we need to decrease the HP length when a pointer is
deactivated.

Everything now works as it should; hazard_ptrs are ready for action!