Added hazard_queue testing.
Testing uncovered a bug in its implementation. This was fixed. Now
the hazard_queue is working smoothly (the bug was with the queue, not
the pointers).

There's no memory leak with hazard_queue, though there is apparantely
some difficulty with the shared_queue. I'll check on this in a different
commit.